@@ -9,15 +9,21 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@rpldy/native-uploady": "^1.4.1",
+    "expo-document-picker": "~11.2.2",
+    "react-native-document-picker": "^8.1.4",
+    "react-native-paper": "^5.6.0",
+    "react-native-safe-area-context": "^4.5.0",
+    "react-native-vector-icons": "^9.2.0",
     "@expo/webpack-config": "^18.0.4",
     "axios": "^1.3.5",
     "expo": "^48.0.11",
     "expo-av": "^13.2.1",
     "expo-status-bar": "~1.4.4",
     "firebase": "^9.19.1",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
-    "react-native": "^0.69.9",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-native": "^0.71.6",
     "react-native-web": "~0.18.10",
     "react-navigation": "^4.4.4",
     "react-navigation-stack": "^2.10.4"
